+++ bigneuron_ported/qul/anisodiffusion/q_EigenVectors3D.h
@@ -10,12 +10,12 @@
 //hessian=|Ixy Iyy Iyz|
 //        |Ixz Iyz Izz|
 //eigenvalue: eigval3>eigval2>eigval1
-bool q_eigenvectors3D(double *p_Ixx,double *p_Ixy,double *p_Ixz,double *p_Iyy,double *p_Iyz,double *p_Izz,
+bool q_eigenvectors3D(float *p_Ixx,float *p_Ixy,float *p_Ixz,float *p_Iyy,float *p_Iyz,float *p_Izz,
 					  long sz_img[4],
-					  double *&p_eigval3,double *&p_eigval2,double *&p_eigval1,
-					  double *&p_eigvec3x,double *&p_eigvec3y,double *&p_eigvec3z,
-					  double *&p_eigvec2x,double *&p_eigvec2y,double *&p_eigvec2z,
-					  double *&p_eigvec1x,double *&p_eigvec1y,double *&p_eigvec1z);
+					  float *&p_eigval3,float *&p_eigval2,float *&p_eigval1,
+					  float *&p_eigvec3x,float *&p_eigvec3y,float *&p_eigvec3z,
+					  float *&p_eigvec2x,float *&p_eigvec2y,float *&p_eigvec2z,
+					  float *&p_eigvec1x,float *&p_eigvec1y,float *&p_eigvec1z);
 
 
 //consider to add mask image to save time - todo
